examples: Consolidate agent.py and weather_agent.py into a single ultimate agent.py example
